This is a pretty big commit. More below
I modifed search engine to not hold function data in memory and whatnot.
I also modified it to query the database based on the idx id that is
returned from the search function of search engine.

Also, I added the ability for the server to serve the static html file
so instead of having to load the file in the browser, it automatically
serves it to you when you enter in http://<ip address>:<PORT>. ez.