Big boi changes
A restructuring of our classes works wonders. I mean, now everything is
nice, need, and abstracted to an easy to use layer (as long as
dependencies are working correctly). Furthermore, I added in the
endpoint for the eventual file adding funtionality that we're hoping
for. Very cool. Stay tuned, because a biggin' will be coming our way
very shortly. EXCITING!